Avoid crash when editing Git commit message
I ran into this issue and discovered that it had already been reported:
https://github.com/TheSpyder/SyncedSideBar/issues/61

This isn't an issue _caused_ by SyncedSideBar - even without the package
installed, if you're editing `.git/COMMIT_EDITMSG` and you right-click
inside the editor and select "Reveal in Side Bar", Sublime Text will
likely crash. SyncedSideBar just triggers the issue when opening the
file, instead of requiring you to manually trigger it.

I've tested this in Sublime Text's safe mode and can reproduce the crash
there, but not reliably. I'll report it as a bug directly to Sublime
Text, hopefully after working out a reliable reproduction case.

This patch works around the issue by avoiding running "Reveal in Side
Bar" if the currently-active view is a file inside a `.git` directory.

This works but isn't a comprehensive solution, and may not be desirable
for everyone. Ideally this would check against Sublime Text's existing
preference lists of excluded files and folders (since those won't be
visible in the side bar anyway), but Sublime Text doesn't seem to expose
an API method for checking files against the exclude list, so that would
require implementing the logic to match file paths against those lists
within this package, which is further than I'm willing to go at the
moment. This workaround avoids the crash, and hopefully the crash will
get fixed in Sublime Text itself soon and I can retire the workaround.