Add version check to Git crash workaround
This has been marked as fixed in Sublime Text 4148:
https://github.com/sublimehq/sublime_text/issues/5881

I'm not sure which version of Sublime Text introduced the crash, so I
haven't added a lower version to the check. This means that
SyncedSideBar won't reveal files under `.git/` in older versions of
Sublime Text, even if there's no risk of a crash in those versions.